edit fabfile

+++ fabfile.py
@@ -21,7 +21,7 @@ from fabric.contrib.files import exists
 
 # -------- fab设置 -------- #
 SSH_NEW_PORT = 22  # 50683                               # rule.v4中开放的端口需跟此一致
-env.hosts = ['10.49.196.52:%d' % SSH_NEW_PORT]           # 如果有多个主机，fabric会自动依次部署
+env.hosts = ['10.49.192.82:%d' % SSH_NEW_PORT]           # 如果有多个主机，fabric会自动依次部署
 env.user = 'luxf'
 # env.hosts = ['206.189.216.83:%d' % SSH_NEW_PORT]           # 如果有多个主机，fabric会自动依次部署
 # env.user = 'root'
@@ -63,12 +63,12 @@ def _prepare():
 
     # download the current stable release of Docker Compose
     # sudo('pip install docker-compose')
-    sudo('curl -L "https://github.com/docker/compose/releases/download/1.23.2/docker-compose-$(uname -s)-$(uname -m)" -o /usr/local/bin/docker-compose')
-    sudo('chmod +x /usr/local/bin/docker-compose')
-    sudo('ln -s /usr/local/bin/docker-compose /usr/bin/docker-compose')
+    if not exists('/usr/local/bin/docker-compose'):
+        sudo('curl -L "https://github.com/docker/compose/releases/download/1.23.2/docker-compose-$(uname -s)-$(uname -m)" -o /usr/local/bin/docker-compose')
+        sudo('chmod +x /usr/local/bin/docker-compose')
+        sudo('ln -s /usr/local/bin/docker-compose /usr/bin/docker-compose')
 
-    # sudo('docker login', pty=False, combine_stderr=True)
-    
+    sudo('docker login')
 
 
 def _ssh_setting():
@@ -105,6 +105,9 @@ def _security_setting():
         sudo('sudo service netfilter-persistent reload')
     except BaseException as e:
         sudo('sudo service iptables-persistent reload')
+
+    sudo('service docker restart')
+    sudo('iptables-save > /etc/iptables/rules.v4')
 
 
 def _get_backup_to_local():
@@ -162,9 +165,8 @@ def init_env():
     #     pass
 
     _prepare()
-    # _security_setting()
-    # sudo('service docker restart')
-    # sudo('iptables-save > /etc/iptables/rules.v4')
+
+    _security_setting()
 
 
 def upload():
@@ -188,7 +190,7 @@ def start():
 def restart():
     """重启 服务"""
     with cd(_REMOTE_DIR_APP):
-        sudo("./deploy.sh reload")
+        sudo("./deploy.sh restart")
 
 
 def stop():
